# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification - Transporte Escolar Eventos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /veiculos:
    get:
      operationId: listVeiculos
      tags: [veiculos]
      summary: Listar veículos
      parameters:
        - name: empresa_id
          in: query
          schema: { type: integer }
        - name: situacao
          in: query
          schema: { type: string }
        - name: tipo_utilizacao
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Lista de veículos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Veiculo"
    post:
      operationId: createVeiculo
      tags: [veiculos]
      summary: Cadastrar veículo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VeiculoInput"
      responses:
        "201":
          description: Veículo criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Veiculo"
  /veiculos/import:
    post:
      operationId: importVeiculos
      tags: [veiculos]
      summary: Importar veículos via planilha
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportRequest"
      responses:
        "200":
          description: Resultado da importação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /veiculos/{id}:
    get:
      operationId: getVeiculo
      tags: [veiculos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Veículo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Veiculo"
    patch:
      operationId: updateVeiculo
      tags: [veiculos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VeiculoUpdate"
      responses:
        "200":
          description: Veículo atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Veiculo"
    delete:
      operationId: deleteVeiculo
      tags: [veiculos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deletado

  # ── Locais de Eventos ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Veiculo:
      type: object
      required: [id, modelo, placa, lugares, empresa_id]
      properties:
        id:              { type: integer }
        prefixo:         { type: ["string", "null"] }
        tipo:            { type: ["string", "null"] }
        marca:           { type: ["string", "null"] }
        modelo:          { type: string }
        placa:           { type: string }
        cor:             { type: ["string", "null"] }
        empresa_id:      { type: integer }
        empresa_nome:    { type: ["string", "null"] }
        lugares:         { type: integer }
        situacao:        { type: ["string", "null"] }
        tipo_utilizacao: { type: ["string", "null"] }
        ano:             { type: ["string", "null"] }
        acessibilidade:  { type: boolean }
    VeiculoInput:
      type: object
      required: [modelo, placa, lugares, empresa_id]
      properties:
        prefixo:         { type: string }
        tipo:            { type: string }
        marca:           { type: string }
        modelo:          { type: string }
        placa:           { type: string }
        cor:             { type: string }
        empresa_id:      { type: integer }
        lugares:         { type: integer }
        situacao:        { type: string }
        tipo_utilizacao: { type: string }
        ano:             { type: string }
        acessibilidade:  { type: boolean }
    VeiculoUpdate:
      type: object
      properties:
        modelo:          { type: string }
        placa:           { type: string }
        cor:             { type: string }
        empresa_id:      { type: integer }
        lugares:         { type: integer }
        situacao:        { type: string }
        tipo_utilizacao: { type: string }
        acessibilidade:  { type: boolean }

    # ── Local ─────────────────────────────────────────────────────────────────
...
    ImportRequest:
      type: object
      required: [fileBase64]
      properties:
        fileBase64: { type: string }
    ImportResult:
      type: object
      required: [total, importados, ignorados, erros]
      properties:
        total:     { type: integer }
        importados: { type: integer }
        ignorados:  { type: integer }
        erros:
          type: array
          items:
            type: string